Replace object storage config with mimir's
