--- pkg/objstore/bucket/filesystem/bucket_client.go
@@ -1,16 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-// Provenance-includes-location: https://github.com/cortexproject/cortex/blob/master/pkg/storage/bucket/filesystem/bucket_client.go
-// Provenance-includes-license: Apache-2.0
-// Provenance-includes-copyright: The Cortex Authors.
-
-package filesystem
-
-import (
-	"github.com/thanos-io/objstore"
-	"github.com/thanos-io/objstore/providers/filesystem"
-)
-
-// NewBucketClient creates a new filesystem bucket client
-func NewBucketClient(cfg Config) (objstore.Bucket, error) {
-	return filesystem.NewBucket(cfg.Directory)
-}
