+++ pkg/objstore/bucket_util_test.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
 
-package bucket
+package objstore
 
 import (
 	"context"
