+++ pkg/objstore/client/client_mock.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
 
-package bucket
+package client
 
 import (
 	"bytes"
