+++ pkg/objstore/delayed_bucket_client.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package bucket
+package objstore
 
 import (
 	"context"
