+++ pkg/objstore/providers/azure/bucket_client.go
