+++ pkg/objstore/providers/gcs/bucket_client.go
