+++ pkg/objstore/providers/s3/config_test.go
