+++ pkg/objstore/providers/swift/bucket_client.go
