+++ pkg/objstore/providers/swift/config.go
